+++ babel.config.js
@@ -1,0 +1,20 @@
+// @ts-check
+
+/** @type {import('@babel/core').TransformOptions} */
+const config = {
+  presets: [
+    [
+      '@babel/preset-env',
+      {
+        modules: false,
+        useBuiltIns: 'entry',
+        corejs: 3,
+      },
+    ],
+    '@babel/preset-typescript',
+    '@babel/preset-react',
+  ],
+  plugins: ['@babel/plugin-syntax-dynamic-import', 'babel-plugin-lodash'],
+}
+
+module.exports = config