+++ cmd/repo-updater/repos/sources.go
@@ -61,15 +61,6 @@ func NewSourcer(cf *httpcli.Factory, decs ...func(Source) Source) Sourcer {
 			srcs = append(srcs, src)
 		}
 
-		if !includesGitHubDotComSource(srcs) {
-			// add a GitHub.com source by default, to support navigating to URL
-			// paths like /github.com/foo/bar to auto-add that repository. This
-			// source returns nothing for ListRepos. However, in the future we
-			// intend to use it in repoLookup.
-			src, err := NewGithubDotComSource(cf)
-			srcs, errs = append(srcs, src), multierror.Append(errs, err)
-		}
-
 		return srcs, errs.ErrorOrNil()
 	}
 }
@@ -94,21 +85,6 @@ func NewSource(svc *ExternalService, cf *httpcli.Factory) (Source, error) {
 	default:
 		panic(fmt.Sprintf("source not implemented for external service kind %q", svc.Kind))
 	}
-}
-
-func includesGitHubDotComSource(srcs []Source) bool {
-	for _, svc := range Sources(srcs).ExternalServices() {
-		if !strings.EqualFold(svc.Kind, "GITHUB") {
-			continue
-		} else if cfg, err := svc.Configuration(); err != nil {
-			continue
-		} else if u, err := url.Parse(cfg.(*schema.GitHubConnection).Url); err != nil {
-			continue
-		} else if strings.HasSuffix(u.Hostname(), "github.com") {
-			return true
-		}
-	}
-	return false
 }
 
 // sourceTimeout is the default timeout to use on Source.ListRepos
@@ -266,6 +242,16 @@ func (s GithubSource) ExternalServices() ExternalServices {
 	return ExternalServices{s.svc}
 }
 
+// GetRepo returns the Github repository with the given name and owner
+// ("org/repo-name")
+func (s GithubSource) GetRepo(ctx context.Context, nameWithOwner string) (*Repo, error) {
+	r, err := s.conn.getRepository(ctx, nameWithOwner)
+	if err != nil {
+		return nil, err
+	}
+	return githubRepoToRepo(s.svc, r, s.conn), nil
+}
+
 func githubRepoToRepo(
 	svc *ExternalService,
 	ghrepo *github.Repository,
@@ -274,6 +260,7 @@ func githubRepoToRepo(
 	urn := svc.URN()
 	return &Repo{
 		Name:         string(githubRepositoryToRepoPath(conn, ghrepo)),
+		URI:          string(reposource.GitHubRepoName("", conn.originalHostname, ghrepo.NameWithOwner)),
 		ExternalRepo: *github.ExternalRepoSpec(ghrepo, *conn.baseURL),
 		Description:  ghrepo.Description,
 		Fork:         ghrepo.IsFork,
@@ -336,6 +323,7 @@ func gitlabProjectToRepo(
 	urn := svc.URN()
 	return &Repo{
 		Name:         string(gitlabProjectToRepoPath(conn, proj)),
+		URI:          string(reposource.GitLabRepoName("", conn.baseURL.Hostname(), proj.PathWithNamespace)),
 		ExternalRepo: *gitlab.ExternalRepoSpec(proj, *conn.baseURL),
 		Description:  proj.Description,
 		Fork:         proj.ForkedFromProject != nil,
@@ -399,6 +387,7 @@ func bitbucketserverRepoToRepo(
 	urn := svc.URN()
 	return &Repo{
 		Name:         string(info.Name),
+		URI:          bitbucketServerRepoURI(conn.config, repo),
 		ExternalRepo: *info.ExternalRepo,
 		Description:  info.Description,
 		Fork:         info.Fork,
@@ -412,6 +401,27 @@ func bitbucketserverRepoToRepo(
 		},
 		Metadata: repo,
 	}
+}
+
+func bitbucketServerRepoURI(config *schema.BitbucketServerConnection, repo *bitbucketserver.Repo) string {
+	// This code follows the same path we generate for Name, except it uses
+	// the default repository pattern.
+	//
+	// TODO(keegancsmith) Cleanup bitbucket code such that we don't need to
+	// constantly parse and normalize the host.
+	host, err := url.Parse(config.Url)
+	if err != nil {
+		// This should never happen
+		panic(fmt.Sprintf("malformed bitbucket config, invalid URL: url=%q error=%s", config.Url, err))
+	}
+	host = NormalizeBaseURL(host)
+
+	project := "UNKNOWN"
+	if repo.Project != nil {
+		project = repo.Project.Key
+	}
+
+	return string(reposource.BitbucketServerRepoName("", host.Hostname(), project, repo.Slug))
 }
 
 // A GitoliteSource yields repositories from a single Gitolite connection configured
@@ -503,8 +513,10 @@ func gitoliteRepoToRepo(
 	conn *schema.GitoliteConnection,
 ) *Repo {
 	urn := svc.URN()
+	name := string(reposource.GitoliteRepoName(conn.Prefix, repo.Name))
 	return &Repo{
-		Name:         string(reposource.GitoliteRepoName(conn.Prefix, repo.Name)),
+		Name:         name,
+		URI:          name,
 		ExternalRepo: *gitolite.ExternalRepoSpec(repo, gitolite.ServiceID(conn.Host)),
 		Enabled:      true,
 		Sources: map[string]*SourceInfo{
@@ -643,6 +655,7 @@ func phabricatorRepoToRepo(
 
 	return &Repo{
 		Name: name,
+		URI:  name,
 		ExternalRepo: api.ExternalRepoSpec{
 			ID:          repo.PHID,
 			ServiceType: "phabricator",
@@ -740,6 +753,7 @@ func awsCodeCommitRepoToRepo(
 
 	awsRepo := &Repo{
 		Name:         string(awsCodeCommitRepositoryToRepoPath(conn, repo)),
+		URI:          string(reposource.AWSRepoName("", repo.Name)),
 		ExternalRepo: *awscodecommit.ExternalRepoSpec(repo, serviceID),
 		Description:  repo.Description,
 		Enabled:      true,
@@ -844,6 +858,7 @@ func otherRepoFromCloneURL(urn string, u *url.URL) *Repo {
 
 	return &Repo{
 		Name: repoName,
+		URI:  repoName,
 		ExternalRepo: api.ExternalRepoSpec{
 			ID:          string(repoName),
 			ServiceType: "other",