+++ migrations/1528395579_.down.sql
@@ -1,0 +1,1 @@
+DROP INDEX IF EXISTS repo_uri_idx;