+++ web/babel.config.js
@@ -2,18 +2,7 @@
 
 /** @type {import('@babel/core').TransformOptions} */
 const config = {
-  plugins: ['@babel/plugin-syntax-dynamic-import', 'babel-plugin-lodash'],
-  presets: [
-    [
-      '@babel/preset-env',
-      {
-        modules: false,
-        targets: require('../package.json').browserslist,
-        useBuiltIns: 'entry',
-        corejs: 3,
-      },
-    ],
-  ],
+  extends: '../babel.config.js',
 }
 
 module.exports = config